quotes_project/pyproject.toml
[tool.poetry]
name = "quotes-project"
version = "0.1.0"
description = ""
authors = ["Halyna Marchenko <marchenkohalyna888@gmail.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
django = ">=5.2.3,<6.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
pymongo = ">=4.13.2,<5.0.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
